Create main page feature.
Create a 'fillers' folder with events.js. This file will contain all
the page loading events.

Create 'main.handlebars' with a h1 section to test that our handlebars
can connect with events.js.

Add fillers.onPageLoad to the document load function so that when the page
initially loads, the handlebars/fillers content will be seen.

@quidprocrow
  @smb2255
    @briankeegan