Add user API file
Add file to make API requests for users.

To do this, our team had to hard code in the API's development URL
on the localhost, despite previously setting the API development
URL in the config file on the browser repo. This is unexpected,
but we are nonetheless able to get a successful sign in
using the hardcoded API url.

@quidprocrow
  @briankeegan
    @smb2255