Create 'create item' feature on profile
Add div to profile handlebars for the create item form

Adjust create item filler event so that it targets the div on the
profile handlebars

Adjust the index items success so that, upon success, the user is prompted
to create an item if they don't have any; if they do, populates as
usual. In addition, clear the create item form, and makes sure that
the bucketListBucket is empty (so that if someone deletes their last
item, it is still cleared).

@quidprocrow
@briankeegan
@smb2255